Create DonationPaymentSession model

+++ database/migrations/2021_08_02_174837_create_donation_payment_sessions.php
@@ -17,7 +17,7 @@ class CreateDonationPaymentSessions extends Migration
             $table->bigIncrements('donation_payment_session_id');
             $table->integer('account_id')->unsigned()->nullable();
             $table->integer('donation_tier_id')->unsigned();
-            $table->integer('donation_perks_id')->unsigned()->nullable();
+            $table->bigInteger('donation_perks_id')->unsigned()->nullable();
             $table->string('stripe_session_id');
             $table->string('stripe_price_id');
             $table->integer('number_of_months')->unsigned()->comment('Purchase quantity');
